requirements.txt
flask
google-genereativai
numpy
pandas
matplotlib
dotenv
groq
google-genai
tabulate